Restructured folders, using data folder for tests
