+++ package.json
@@ -7,8 +7,8 @@
   "scripts": {
     "test": "./node_modules/.bin/nyc -e 'ts' -r lcov -r text yarn run unitTests",
     "prepublish": "rm -rf ./js/; ./node_modules/.bin/tsc -p ./tsconfig.json",
-    "unitTests": "mocha tests/*.test.ts -r ts-node/register",
-    "test:watch": "mocha  --watch-extensions ts tests/*.test.ts -r ts-node/register -w"
+    "unitTests": "mocha tests/**/*.test.ts -r ts-node/register",
+    "test:watch": "mocha  --watch-extensions ts tests/**/*.test.ts -r ts-node/register -w"
   },
   "repository": {
     "type": "git",
